+++ src/bpmncwpverify/core/frontends/cwpXmlParser.py
