+++ test/core/test_bpmn.py
@@ -4,9 +4,9 @@ from xml.etree.ElementTree import Element, SubElement, tostring
 from defusedxml import ElementTree
 from returns.result import Failure, Success
 
-from bpmncwpverify.core.accessmethods.bpmnmethods import from_xml
 from bpmncwpverify.core.bpmn import BPMN_XML_NAMESPACE, StartEvent
 from bpmncwpverify.core.error import BpmnMissingEventsError
+from bpmncwpverify.core.frontends.bpmnmethods import from_xml
 from bpmncwpverify.core.state import StateBuilder
