+++ test/core/test_cwp.py
@@ -5,7 +5,6 @@ import pytest
 from returns.functions import not_
 from returns.pipeline import is_successful
 
-from bpmncwpverify.core.accessmethods.cwpmethods import CwpXmlParser
 from bpmncwpverify.core.cwp import CwpEdge, CwpState
 from bpmncwpverify.core.error import (
     CwpMultStartStateError,
@@ -13,6 +12,7 @@ from bpmncwpverify.core.error import (
     CwpNoStartStateError,
     Error,
 )
+from bpmncwpverify.core.frontends.cwpmethods import CwpXmlParser
 from bpmncwpverify.core.state import State
